+++ src/main/java/com/ayaan/FinanceTracker/dao/IncomeDAO.java
@@ -2,102 +2,21 @@ package com.ayaan.FinanceTracker.dao;
 
 import java.util.List;
 
-import org.hibernate.Session;
-import org.hibernate.Transaction;
-
 import com.ayaan.FinanceTracker.models.Income;
-import com.ayaan.FinanceTracker.util.HibernateUtil;
 
-import jakarta.persistence.NoResultException;
-import jakarta.persistence.Query;
+public interface IncomeDAO {
 
-public class IncomeDAO {
+    void updateIncome(Income income);
 
-    public void saveIncome(Income income) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.save(income);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
+    void deleteIncome(Income income);
 
-    public void updateIncome(Income income) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.update(income);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
+    Income getIncomebyId(Integer id);
 
-    public void deleteIncome(Income income) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            // BankAccount bankAccount = session.get(BankAccount.class, id);
-            if (income != null) {
-                session.delete(income);
-                System.out.println("Bank Account is deleted");
-            }
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
+    List<Income> getAllIncome();
 
-    public Income getIncomebyId(Integer id) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.get(Income.class, id);
-        }
-    }
+    Double getIncomes();
 
-    public List<Income> getAllIncome() {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from Income", Income.class).list();
-        }
-    }
+    Double getIncomeBySourceFromIncomes(String source);
 
-    public Double getIncomes() {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            String hql = "SELECT SUM(i.income) FROM Income i";
-            return session.createQuery(hql, Double.class).uniqueResult(); 
-        }
-    }
-
-    public Double getIncomeBySourceFromIncomes(String source) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            String hql = "SELECT i.income FROM Income i " +
-                         "JOIN i.incomeSources ies " +
-                         "WHERE ies.incomeExpenseSource = :source";
-            Query query = session.createQuery(hql);
-            query.setParameter("source", source);
-
-            try {
-                return (Double) query.getSingleResult();
-                
-            } catch (NoResultException e) {
-                System.out.println("No income found for the given source.");
-                return null;
-
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-    
+    void saveIncome(Income income);
 }
